manual rollback of interface name
