+++ backend/src/interface/mod.rs
@@ -1,7 +1,7 @@
 //! Module bundle for interface models
 pub mod admin_notification;
 pub mod api_command;
-pub mod image_api;
+pub mod image_validation;
 pub mod image_storage;
 pub mod mealplan_management;
 pub mod mensa_parser;
