--- backend/src/layer/data/image_api/image_verification.rs
@@ -1,15 +1,0 @@
-use crate::interface::image_api::Result;
-
-pub struct ImageVerification {
-    acceptance: [u8; 5],
-}
-
-impl ImageVerification {
-    fn new(acceptance: [u8; 5]) -> Self {
-        Self { acceptance }
-    }
-
-    fn verify() -> Result<()> {
-        todo!()
-    }
-}
