+++ backend/src/layer/data/image_validation/mod.rs
@@ -1,6 +1,6 @@
 //! This component is responsible for image related api operations
 pub mod api_request;
 pub mod google_api_handler;
-pub mod image_verification;
+pub mod image_evaluation;
 pub mod json_parser;
 pub mod json_structs;
