+++ backend/src/layer/data/mod.rs
@@ -1,6 +1,6 @@
 //! In this layer all interactions with the outside world and the required conversions of data take place.
 
 pub mod database;
-pub mod image_api;
+pub mod image_validation;
 pub mod mail;
 pub mod swka_parser;
